// src/components/types/form.ts
interface FormInputType {
    text: string
    email: string
    select: string
    textArea: string
    file: string
    checkBox: string
    date: string
    password: string
}

export const formInputType: FormInputType = {
    text: "text",
    email: "email",
    select: "select",
    textArea: "textArea",
    file: "file",
    checkBox: "checkbox",
    date: "date",
    password: "password"
}